+++ app/Http/Controllers/DocumentController.php
@@ -4,10 +4,10 @@ namespace App\Http\Controllers;
 
 use App\Corpus;
 use App\Document;
-use App\Association;
 use App\Custom\GitRepoInterface;
 use App\Custom\LaudatioUtilsInterface;
 use GrahamCampbell\Flysystem\FlysystemManager;
+use App\Custom\ElasticsearchInterface;
 use Illuminate\Http\Request;
 use Response;
 use Log;
@@ -18,13 +18,15 @@ class DocumentController extends Controller
 
     protected $GitRepoService;
     protected $LaudatioUtilService;
+    protected $elasticService;
     protected $flysystem;
 
-    public function __construct(GitRepoInterface $GitRepoService, FlysystemManager $flysystem,LaudatioUtilsInterface $laudatioUtils)
+    public function __construct(GitRepoInterface $GitRepoService, FlysystemManager $flysystem,LaudatioUtilsInterface $laudatioUtils,ElasticsearchInterface $elasticService)
     {
         $this->GitRepoService = $GitRepoService;
         $this->LaudatioUtilService = $laudatioUtils;
         $this->flysystem = $flysystem;
+        $this->elasticService = $elasticService;
     }
 
     /**
@@ -136,10 +138,11 @@ class DocumentController extends Controller
 
     public function destroyDocumentContent(Request $request) {
         $result = array();
-        $status = "";
+        $status = "success";
 
         try{
             $corpusid = $request->input('corpusid');
+            $corpus = Corpus::findOrFail($corpusid);
             $corpusPath = $request->input('path');
             $auth_user_name = $request->input('auth_user_name');
             $auth_user_id = $request->input('auth_user_id');
@@ -147,24 +150,39 @@ class DocumentController extends Controller
             $toBeDeletedCollection = $request->input('tobedeleted');
 
             foreach ($toBeDeletedCollection as $toBeDeleted) {
-                $this->GitRepoService->deleteFile($this->flysystem,$corpusPath."/".$toBeDeleted['fileName'],$auth_user_name,$auth_user_email);
-                $document = Document::findOrFail($toBeDeleted['databaseId']);
+                $fileDeleteResult = $this->GitRepoService->deleteFile($this->flysystem,$corpusPath."/".$toBeDeleted['fileName'],$auth_user_name,$auth_user_email);
 
-                if(count($document->annotations()) > 0) {
-                    $document->annotations()->detach();
-                }
-                $document->delete();
+                if($fileDeleteResult) {
+                    $document = Document::findOrFail($toBeDeleted['databaseId']);
+
+                    $searchData = array();
+                    array_push($searchData,array(
+                        "_id" => $document->elasticsearch_id
+                    ));
+
+                    $elasticSearchDeleteResult = $this->elasticService->deleteIndexedObject($document->elasticsearch_index,$searchData);
+
+                    if(!$elasticSearchDeleteResult['error']) {
+                        if(count($document->annotations()) > 0) {
+                            $document->annotations()->detach();
+                        }
+                        $document->delete();
+
+                        $pushResult = $this->GitRepoService->pushFiles($corpusPath,$corpusid,$auth_user_name);
+                    }
+
+                    $this->LaudatioUtilService->emptyDocumentCacheByCorpusId($corpus->corpus_id,$document->elasticsearch_index);
+                    $this->LaudatioUtilService->emptyDocumentCacheByDocumentIndex($document->elasticsearch_index);
+                }//end if filedeleted
             }
 
-            $this->GitRepoService->pushFiles($corpusPath,$corpusid,$auth_user_name);
 
-            $status = "success";
-            $result['delete_document_content_response']  = "Document content was successfully deleted";
+            $result['delete_content_response']  = "Document content was successfully deleted";
 
         }
         catch (\Exception $e) {
             $status = "error";
-            $result['delete_document_content_response']  = "There was a problem deleting the Document content. The error was: ($e) A message has been sent to the site administrator. Please try again later";
+            $result['delete_content_response']  = "There was a problem deleting the Document content. The error was: ($e) A message has been sent to the site administrator. Please try again later";
         }
